Jordan Changes
did a lot of revamping of various files.
Completed and tested displayAssignment functions
completed Evan's ReadAssignmentsFromFile
and renamed it to importAssignmentsFromFile.

the function now adds assignments to the assignedAssignments list. We
need clarification on whether we will be reading in completed and
assigned assignments FROM file or if they will all be assigned
assignments. If it is the former, I will make appropriate changes since
I am familiar with how these functions are working